service: set: Refractor setting service
