+++ src/core/hle/service/set/setting_formats/appln_settings.cpp
@@ -1,11 +1,11 @@
 // SPDX-FileCopyrightText: Copyright 2018 yuzu Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
 
-#include "core/hle/service/set/private_settings.h"
+#include "core/hle/service/set/setting_formats/appln_settings.h"
 
 namespace Service::Set {
 
-PrivateSettings DefaultPrivateSettings() {
+ApplnSettings DefaultApplnSettings() {
     return {};
 }
