+++ src/core/hle/service/set/setting_formats/device_settings.cpp
@@ -1,11 +1,11 @@
 // SPDX-FileCopyrightText: Copyright 2018 yuzu Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
 
-#include "core/hle/service/set/appln_settings.h"
+#include "core/hle/service/set/setting_formats/device_settings.h"
 
 namespace Service::Set {
 
-ApplnSettings DefaultApplnSettings() {
+DeviceSettings DefaultDeviceSettings() {
     return {};
 }
