+++ src/core/hle/service/set/setting_formats/device_settings.h
@@ -7,10 +7,12 @@
 #include <cstddef>
 
 #include "common/common_types.h"
+#include "common/vector_math.h"
+#include "core/hle/service/set/settings_types.h"
 
 namespace Service::Set {
 struct DeviceSettings {
-    std::array<u8, 0x10> reserved_000;
+    INSERT_PADDING_BYTES(0x10); // Reserved
 
     // nn::settings::BatteryLot
     std::array<u8, 0x18> ptm_battery_lot;
@@ -19,26 +21,24 @@ struct DeviceSettings {
     u8 ptm_battery_version;
     // nn::settings::system::PtmCycleCountReliability
     u32 ptm_cycle_count_reliability;
-
-    std::array<u8, 0x48> reserved_048;
+    INSERT_PADDING_BYTES(0x48); // Reserved
 
     // nn::settings::system::AnalogStickUserCalibration L
     std::array<u8, 0x10> analog_user_stick_calibration_l;
     // nn::settings::system::AnalogStickUserCalibration R
     std::array<u8, 0x10> analog_user_stick_calibration_r;
-
-    std::array<u8, 0x20> reserved_0B0;
+    INSERT_PADDING_BYTES(0x20); // Reserved
 
     // nn::settings::system::ConsoleSixAxisSensorAccelerationBias
-    std::array<u8, 0xC> console_six_axis_sensor_acceleration_bias;
+    Common::Vec3<f32> console_six_axis_sensor_acceleration_bias;
     // nn::settings::system::ConsoleSixAxisSensorAngularVelocityBias
-    std::array<u8, 0xC> console_six_axis_sensor_angular_velocity_bias;
+    Common::Vec3<f32> console_six_axis_sensor_angular_velocity_bias;
     // nn::settings::system::ConsoleSixAxisSensorAccelerationGain
     std::array<u8, 0x24> console_six_axis_sensor_acceleration_gain;
     // nn::settings::system::ConsoleSixAxisSensorAngularVelocityGain
     std::array<u8, 0x24> console_six_axis_sensor_angular_velocity_gain;
     // nn::settings::system::ConsoleSixAxisSensorAngularVelocityTimeBias
-    std::array<u8, 0xC> console_six_axis_sensor_angular_velocity_time_bias;
+    Common::Vec3<f32> console_six_axis_sensor_angular_velocity_time_bias;
     // nn::settings::system::ConsoleSixAxisSensorAngularAcceleration
     std::array<u8, 0x24> console_six_axis_sensor_angular_acceleration;
 };
