+++ src/core/hle/service/set/setting_formats/system_settings.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright 2018 yuzu Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
 
-#include "core/hle/service/set/system_settings.h"
+#include "core/hle/service/set/setting_formats/system_settings.h"
 
 namespace Service::Set {
