+++ src/core/hle/service/set/system_settings_server.h
@@ -12,10 +12,11 @@
 #include "common/uuid.h"
 #include "core/hle/result.h"
 #include "core/hle/service/service.h"
-#include "core/hle/service/set/appln_settings.h"
-#include "core/hle/service/set/device_settings.h"
-#include "core/hle/service/set/private_settings.h"
-#include "core/hle/service/set/system_settings.h"
+#include "core/hle/service/set/setting_formats/appln_settings.h"
+#include "core/hle/service/set/setting_formats/device_settings.h"
+#include "core/hle/service/set/setting_formats/private_settings.h"
+#include "core/hle/service/set/setting_formats/system_settings.h"
+#include "core/hle/service/set/settings_types.h"
 #include "core/hle/service/time/clock_types.h"
 #include "core/hle/service/time/time_zone_types.h"
 
@@ -24,25 +25,6 @@ class System;
 }
 
 namespace Service::Set {
-enum class GetFirmwareVersionType {
-    Version1,
-    Version2,
-};
-
-struct FirmwareVersionFormat {
-    u8 major;
-    u8 minor;
-    u8 micro;
-    INSERT_PADDING_BYTES(1);
-    u8 revision_major;
-    u8 revision_minor;
-    INSERT_PADDING_BYTES(2);
-    std::array<char, 0x20> platform;
-    std::array<u8, 0x40> version_hash;
-    std::array<char, 0x18> display_version;
-    std::array<char, 0x80> display_title;
-};
-static_assert(sizeof(FirmwareVersionFormat) == 0x100, "FirmwareVersionFormat is an invalid size");
 
 Result GetFirmwareVersionImpl(FirmwareVersionFormat& out_firmware, Core::System& system,
                               GetFirmwareVersionType type);
